# Optimize Materialized View Status Maintenance for Partitioned Tables