… src/test/regress/sql/matview_data.sql
… src/test/singlenode_regress/expected/matview_data.out
… src/test/singlenode_regress/sql/matview_data.sql