@avamingli 2025-04-08
TPC-B 

**before this commit:** 

| round       | avg          | round 1     | round 2     | round 3     |
|-------------|--------------|-------------|-------------|-------------|
| transactions| 367850       | 368535      | 368258      | 366757      |
| latency     | 16.30733333  | 16.277      | 16.289      | 16.356      |
| TPS         | 1226.43066   | 1228.694187 | 1227.83674  | 1222.761054 |


**after this commit:** 
| round        | avg          | round 1     | round 2     | round 3     |
|--------------|--------------|-------------|-------------|-------------|
| transactions | 370138       | 368501      | 370823      | 371090      |
| latency(ms)  | 16.20633333  | 16.279      | 16.176      | 16.164      |
| TPS          | 1234.093081  | 1228.564701 | 1236.364521 | 1237.350022 |

**Summary**

| Summary      | with PR 990   |
|--------------|----------|
| transactions | +0.62%   |
| latency      | -0.62%   |
| TPS          | +0.62%   |